Allow PHPUnit ^11 and ^12
Allow installing newer versions of PHPUnit.
In this library itself only PHPUnit ^10 is possible because of the PHP version constraint (PHPUnit ^11 requires at least PHP 8.2).
But this PR allows consuming packages to install a newer version of PHPUnit.

PHPUnit 12 requires dataProvider to be attributes, so all dataProvider annotations were converted to attributes.
To be able to install PHPUnit 11, psalm ^6 is required, so it was also bumped and some new issues were fixed.

+++ test/PHPUnit/Controller/AbstractControllerTestCaseTest.php
@@ -16,6 +16,7 @@ use LaminasTest\Test\ExpectedExceptionTrait;
 use org\bovigo\vfs\vfsStream;
 use org\bovigo\vfs\vfsStreamWrapper;
 use PHPUnit\Framework\AssertionFailedError;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\ExpectationFailedException;
 use RuntimeException;
 
@@ -57,7 +58,11 @@ class AbstractControllerTestCaseTest extends AbstractHttpControllerTestCase
 
     public static function rmdir(string $dir): bool
     {
-        $files = array_diff(scandir($dir), ['.', '..']);
+        $files = scandir($dir);
+        if ($files === false) {
+            throw new RuntimeException('Failed to scan directory');
+        }
+        $files = array_diff($files, ['.', '..']);
         foreach ($files as $file) {
             is_dir("$dir/$file") ? static::rmdir("$dir/$file") : unlink("$dir/$file");
         }
@@ -87,7 +92,11 @@ class AbstractControllerTestCaseTest extends AbstractHttpControllerTestCase
     {
         $config = $this->getApplicationConfig();
         $config = $config['module_listener_options']['cache_dir'];
-        $this->assertEquals(0, count(glob($config . '/*.php')));
+        $files  = glob($config . '/*.php');
+        if ($files === false) {
+            throw new RuntimeException('Failed to glob cache directory');
+        }
+        $this->assertEquals(0, count($files));
     }
 
     /** @return void */
@@ -618,6 +627,7 @@ class AbstractControllerTestCaseTest extends AbstractHttpControllerTestCase
 
     /**
      * @psalm-return Generator<string, array{0: null|string}, mixed, void>
+     * @psalm-suppress PossiblyUnusedMethod Psalm does not yet recognize the DataProvider attribute
      */
     public static function method(): Generator
     {
@@ -629,9 +639,7 @@ class AbstractControllerTestCaseTest extends AbstractHttpControllerTestCase
         yield 'patch' => ['PATCH'];
     }
 
-    /**
-     * @dataProvider method
-     */
+    #[DataProvider('method')]
     public function testDispatchWithNullParams(?string $method): void
     {
         $this->dispatch('/custom-response', $method, null);
@@ -674,6 +682,7 @@ class AbstractControllerTestCaseTest extends AbstractHttpControllerTestCase
 
     /**
      * @return iterable<non-empty-string, array{non-empty-string}>
+     * @psalm-suppress PossiblyUnusedMethod Psalm does not yet recognize the DataProvider attribute
      */
     public static function routeParam(): iterable
     {
@@ -681,9 +690,7 @@ class AbstractControllerTestCaseTest extends AbstractHttpControllerTestCase
         yield 'param' => ['param'];
     }
 
-    /**
-     * @dataProvider routeParam
-     */
+    #[DataProvider('routeParam')]
     public function testRequestWithRouteParam(string $param): void
     {
         $this->dispatch(sprintf('/with-param/%s', $param));
